refactor: rename package from mqrest to mqrestadmin
The package name should reflect that this wraps the IBM MQ
administrative REST API specifically, not a general MQ REST client.

- Rename mqrest/ directory to mqrestadmin/
- Update all package declarations
- Update error message prefixes
- Update doc.go package comment and code examples
- Update CLAUDE.md and plan documentation references
