+++ mqrestadmin/auth_test.go
@@ -1,4 +1,4 @@
-package mqrest
+package mqrestadmin
 
 import (
 	"encoding/base64"
