+++ mqrestadmin/mapping_data.go
@@ -1,4 +1,4 @@
-package mqrest
+package mqrestadmin
 
 import _ "embed"
